feat: add product/merchant reviews, nested categories, and rider delivery workflow
- Deliveries: Added `DeliveryController` enabling Riders to view available orders, accept, pick up, and confirm deliveries (OTP). Updated `Order` models and migrations to track these states.
- Reviews: Introduced `ProductReview` and `MerchantReview` models with their respective migrations. Added customer endpoints to submit reviews.
- E-Commerce Catalog: Updated `ProductCategory` with a `parent_id` column to support nested subcategories. Refactored `EcommerceController` to fetch nested categories, detailed store info, and single product views.
- Merchant Profiles: Added location tracking (lat/long or address specifics) to `MerchantProfile`.
- Testing & Seeding: Added factories for `Product` and `MerchantProfile`. Introduced `DummyUserSeeder`, `DummyMerchantSeeder`, and `ProductSeeder` to populate test data effectively.

+++ database/migrations/2026_08_26_054815_create_product_variants_table.php
@@ -23,8 +23,8 @@ return new class extends Migration
             // Decimal price adjustment if this variant costs more/less than base_price
             $table->decimal('price_adjustment', 10, 2)->default(0.00); 
             
-            // Stock quantity as a decimal to support fractional grocery (e.g. 1.5 kg)
-            $table->decimal('stock_quantity', 10, 2)->default(0.00);
+            // Stock quantity as an integer
+            $table->integer('stock_quantity')->default(0);
 
             $table->timestamps();
         });
